third_party/WebKit/LayoutTests/inspector/sources/debugger/source-frame-breakpoint-decorations.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../../../http/tests/inspector/live-edit-test.js"></script>
<script src="resources/edit-me-breakpoints.js"></script>
<script>

function test()
{
    function waitAndDumpDecorations(sourceFrame)
    {
        return InspectorTest.waitJavaScriptSourceFrameBreakpoints(sourceFrame)
            .then(() => InspectorTest.dumpJavaScriptSourceFrameBreakpoints(sourceFrame));
    }

    Bindings.breakpointManager._storage._breakpoints = {};
    InspectorTest.runDebuggerTestSuite([
        function testAddRemoveBreakpoint(next)
        {
            var javaScriptSourceFrame;
            InspectorTest.showScriptSource("edit-me-breakpoints.js", addBreakpoint);

            function addBreakpoint(sourceFrame)
            {
                javaScriptSourceFrame = sourceFrame;              
                InspectorTest.prepareSourceFrameForBreakpointTest(sourceFrame);
                InspectorTest.addResult("Setting breakpoint");
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 2, "", true)
                    .then(() => waitAndDumpDecorations(javaScriptSourceFrame))
                    .then(removeBreakpoint);
            }

            function removeBreakpoint()
            {
                InspectorTest.addResult("Toggle breakpoint");
                InspectorTest.toggleBreakpoint(javaScriptSourceFrame, 2);
                waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
            }
        },

        function testTwoBreakpointsResolvedInOneLine(next)
        {
            var javaScriptSourceFrame;
            InspectorTest.showScriptSource("edit-me-breakpoints.js", addBreakpoint);

            function addBreakpoint(sourceFrame)
            {
                javaScriptSourceFrame = sourceFrame;              
                InspectorTest.prepareSourceFrameForBreakpointTest(sourceFrame);
                InspectorTest.addResult("Setting breakpoint");
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 2, "", true)
                    .then(() => InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 0, "true", true))
                    .then(() => waitAndDumpDecorations(javaScriptSourceFrame))
                    .then(removeBreakpoint);
            }

            function removeBreakpoint()
            {
                InspectorTest.addResult("Toggle breakpoint");
                InspectorTest.toggleBreakpoint(javaScriptSourceFrame, 2);
                waitAndDumpDecorations(javaScriptSourceFrame).then(() => next());
            }
        },

        function testDecorationInGutter(next)
        {
            var javaScriptSourceFrame;
            InspectorTest.showScriptSource("edit-me-breakpoints.js", addRegularDisabled);

            function addRegularDisabled(sourceFrame)
            {
                InspectorTest.prepareSourceFrameForBreakpointTest(sourceFrame);
                javaScriptSourceFrame = sourceFrame;              
                InspectorTest.addResult("Adding regular disabled breakpoint");
                waitAndDumpDecorations(javaScriptSourceFrame).then(addConditionalDisabled);
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 2, "", false);
            }

            function addConditionalDisabled()
            {
                InspectorTest.addResult("Adding conditional disabled breakpoint");
                waitAndDumpDecorations(javaScriptSourceFrame).then(addRegularEnabled);
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 0, "true", false);
            }

            function addRegularEnabled()
            {
                InspectorTest.addResult("Adding regular enabled breakpoint");
                waitAndDumpDecorations(javaScriptSourceFrame).then(addConditionalEnabled);
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 0, "", true);
            }

            function addConditionalEnabled()
            {
                InspectorTest.addResult("Adding conditional enabled breakpoint");
                waitAndDumpDecorations(javaScriptSourceFrame).then(disableAll);
                InspectorTest.createNewBreakpoint(javaScriptSourceFrame, 0, "true", true);                
            }

            function disableAll()
            {
                InspectorTest.addResult("Disable breakpoints");
                waitAndDumpDecorations(javaScriptSourceFrame).then(enabledAll);
                InspectorTest.toggleBreakpoint(javaScriptSourceFrame, 2, true);
            }

            function enabledAll()
            {
                InspectorTest.addResult("Enable breakpoints");
                waitAndDumpDecorations(javaScriptSourceFrame).then(removeAll);
                InspectorTest.toggleBreakpoint(javaScriptSourceFrame, 2, true);
            }

            function removeAll()
            {
                InspectorTest.addResult("Remove breakpoints");
                waitAndDumpDecorations(javaScriptSourceFrame).then(next);
                InspectorTest.toggleBreakpoint(javaScriptSourceFrame, 2, false);                
            }
        }
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Checks that JavaScriptSourceFrame show breakpoints correctly</p>
</body>
</html>
